Change (u)int to qsizetype (size_t) where applicable
Deliberately omitted:
- QML interfacing code
- public APIs
- sensorfw, whose code is disabled/not practical to compile,
  and whose future is uncertain

Fixes: QTBUG-107069

+++ tests/manual/sensorclerk/collector.cpp
@@ -110,7 +110,7 @@ void Collector::startCollecting()
 {
     if (dataFile.exists()) {
         fileCounter++;
-        for (int i = 0; i < fileCounter; i++) {
+        for (size_t i = 0; i < fileCounter; i++) {
             if (!QFileInfo(QString(QDir::tempPath()+"/sensordump_%1.dat").arg(fileCounter)).exists())
                 dataFile.setFileName(QString(QDir::tempPath()+"/sensordump_%1.dat").arg(fileCounter));
             break;
